Create README.md
The theremin is an electronic musical instrument controlled without physical contact by the performer. The instrument’s controlling section is using hands to control the frequency and amplitude of the sound. The project is designed to simulate theremin, using a computer camera to capture the gesture of the hand and calculate the area of the hand, judge the location of hand to control the frequency and amplitude of the sound.